update header elements
updated logo, js-code refactoring
