+++ html/main.js
@@ -1,6 +1,5 @@
 import { loadHTML } from './src/js/common/htmlBundler.js';
 import { makeAdvantagesSectionMarkup } from './src/js/advantages/makeAdvantagesSection.js';
-import { onMenuClick } from './src/js/header/onMenuClick.js';
 import { onHeaderLoad } from './src/js/header/onHeaderLoad.js';
 import { BLOCK_TYPE } from './src/js/common/constants.js';
 import { fetchVideos } from './src/js/portfolio/fetchVideos.js';
@@ -16,11 +15,14 @@ await loadHTML(BLOCK_TYPE.SECTION, './src/partials/blog.html');
 await loadHTML(BLOCK_TYPE.SECTION, './src/partials/faq.html');
 await loadHTML(BLOCK_TYPE.FOOTER, './src/partials/footer.html');
 
+// Header
+onHeaderLoad()
+
 // Section rendering
 makeAdvantagesSectionMarkup('advantages-glr', './src/template/advantage-cards.hbs');
-onHeaderLoad()
-onMenuClick();
 
 // Portfolio
 fetchVideos();
 initializeCategoryToggle();
+
+
