+++ html/src/css/header.css
@@ -21,7 +21,7 @@
 .header-container.isOpened {
   padding: 16px 32px 32px 32px;
   border-radius: 32px;
-  height: auto;
+  /* height: auto; */
 }
 
 .header-navigation-btn {
@@ -43,18 +43,13 @@
   transform: translateY(8px);
 }
 
-.header-logo-thumb {
-  max-height: 36px;
-}
-
 .header-logo {
   display: block;
-  height: 100%;
+  height: 36px;
 }
 
 .header-logo-link {
   display: inline-block;
-  height: 36px;
 }
 
 .header-navigation-list {
@@ -64,6 +59,10 @@
   align-items: center;
   margin-top: 32px;
   margin-bottom: 32px;
+}
+
+.header-container.isOpened .header-navigation-list {
+  margin-top: 24px;
 }
 
 .header-navigation-link {
@@ -121,6 +120,14 @@
   .header-container.isOpened {
     padding: 24px 32px 32px 64px;
     border-radius: 32px;
+  }
+
+  .header-logo-thumb {
+    max-height: 46px;
+  }
+
+  .header-logo {
+    height: 46px;
   }
 
   .header-navigation-btn {
