+++ html/src/js/header/onHeaderLinksClick.js
@@ -1,0 +1,15 @@
+import { smoothMenuOpening } from "./smoothMenuOpening.js";
+import { onClick } from "./onClick.js";
+
+export function onHeaderLinksClick() {
+  const navLinksRef = document.querySelectorAll('.header-navigation-link');
+  const toFormularLink = document.querySelector('.header-formular-link');
+
+  if (!navLinksRef || !toFormularLink) return;
+
+  [...navLinksRef, toFormularLink].forEach(link => {
+    link.addEventListener('click', onClick);
+
+    smoothMenuOpening(link);
+  });
+}
