+++ html/src/js/header/onHeaderMenuBtnClick.js
@@ -1,0 +1,12 @@
+import { onClick } from "./onClick.js";
+import { smoothMenuOpening } from "./smoothMenuOpening.js";
+
+export function onHeaderMenuBtnClick() {
+  const btnRef = document.getElementById('menu-btn');
+
+  if (!btnRef) return;
+
+  btnRef.addEventListener('click', onClick);
+
+  smoothMenuOpening(btnRef);
+}
